// src/app/models/ligneCommandeWeb.ts
export class LigneCommandeWeb {
    NumeroCommandeWeb?: string ;
    DateCommandeWeb?: Date;
    CodeArticle?: string ;
    Designation?: string ;
    NumeroOrdre?: number;
    Imprimer?: boolean;
    PrixVenteHt?: number;
    Quantite?: number;
    MontantHt?: number;
    TauxRemise?: number;
    MontantRemise?: number;
    MontantFodec?: number;
    NetHt?: number;
    TauxTva?: number;
    MontantTva?: number;
    MontantTtc?: number;
    PrixAchatNet?: number;
    Description?: string;
    Etat?: string ;
    CodeTva?: string ;
    NomUtilisateurRetour?: string;
    Grammage?: boolean;
    LiteTauxRemise?: number;
    CodeTypeCarte?: string ;
    NumeroOrdreRelative?: number;

  }