// src/app/models/article.ts
export interface Article {
    codeArticle?: string;
    codeFamille?: string;
    codeSousFamille?: string;
    codeType?: string;
    codeMarque?: string;
    codeUnite?: string;
    codeUniteVente?: string | null;
    valeurUniteVente?: number;
    designation?: string;
    dernierPrixAchatHt?: number;
    dernierTauxRemise?: number;
    prixAchatHt?: number;
    marge?: number;
    prixVenteHt?: number;
    codeTva?: string;
    prixVenteTtc?: number;
    liteTauxRemise?: number;
    fodec?: boolean;
    stockable?: boolean;
    actif?: boolean;
    tailleCouleur?: boolean;
    observation?: string;
    numeroSerie?: boolean;
    codeBarre?: string;
    imageCodeBarre?: Uint8Array | null;
    logo?: Uint8Array | null;
    prixVenteNetHt?: number | null;
    prixVenteTtcsansRemise?: number;
    prixAchatTtcsansRemise?: number;
    prixVenteHtemporter?: number;
    codeTvaemporter?: string;
    prixVenteTtcemporter?: number;
    avecAccompagnement?: boolean;
    vendu?: boolean | null;
    codeDepotVente?: string | null;
    compose?: boolean;
    valeurConversionUniteAchatVente?: number;
    prixAchatTtc?: number;
    venteEnDetails?: boolean;
    valeurUniteAchat?: number;
    codeArticleDeduction?: string;
    grammage?: boolean | null;
    vip?: boolean;
    isExist?:boolean;
    quantiteCommander?:number;
  }